# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  sheetLink: string;
  /** @nullable */
  customPanels?: string | null;
  isDemoUser?: boolean;
  permissions?: string;
}

export interface DemoUser {
  id: number;
  username: string;
  permissions: string;
  isActive: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface CreateDemoUserInput {
  username: string;
  password: string;
  permissions: string;
  isActive: boolean;
  duration: string;
}

export interface UpdateDemoUserInput {
  password?: string;
  permissions?: string;
  isActive?: boolean;
  duration?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RegisterInput {
  username: string;
  email: string;
  password: string;
  sheetLink: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface GeneratorRecord {
  id: number;
  tDate: string;
  generatorId: string;
  status: string;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  hours?: number | null;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  deliveryStatus?: string | null;
  /** @nullable */
  deliveryTo?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface GeneratorInput {
  tDate: string;
  generatorId: string;
  status: string;
  rating?: string;
  hours?: number;
  remarks?: string;
}

export interface GeneratorUpdate {
  tDate?: string;
  generatorId?: string;
  status?: string;
  rating?: string;
  hours?: number;
  remarks?: string;
  deliveryStatus?: string;
  /** @nullable */
  deliveryTo?: string | null;
}

export type GeneratorStatsByStatusItem = {
  status: string;
  count: number;
};

export interface GeneratorStats {
  total: number;
  byStatus: GeneratorStatsByStatusItem[];
  /** @nullable */
  avgHours?: number | null;
  recentCount: number;
  currentDelivery: number;
  previousDelivery: number;
}

export type UpdateMeBody = {
  customPanels?: string;
};

export type ListGeneratorsParams = {
search?: string;
status?: string;
generatorId?: string;
dateFrom?: string;
dateTo?: string;
};