# lib/api-zod/src/generated/types/demoUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DemoUser {
  id: number;
  username: string;
  permissions: string;
  isActive: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}