# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  sheetLink: string;
  /** @nullable */
  customPanels?: string | null;
  isDemoUser?: boolean;
  permissions?: string;
}